# book-network/src/main/java/com/yassineProject/book_network/auth/AuthenticationResponse.java
package com.yassineProject.book_network.auth;

import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class AuthenticationResponse {
    private String token;
}