# book-network/src/main/java/com/yassineProject/book_network/auth/RegistrationRequest.java
package com.yassineProject.book_network.auth;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.Size;
import lombok.Builder;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
@Builder
public class RegistrationRequest {
    @NotEmpty(message="FirstName is mandatory")
    @NotBlank(message="FirstName is mandatory")
    private String firstName;
    @NotEmpty(message="LastName is mandatory")
    @NotBlank(message="LastName is mandatory")
    private String lastName;
    @NotEmpty(message="Email is mandatory")
    @NotBlank(message="Email is mandatory")
    @Email(message="Email is not formatted")
    private String email;
    @NotEmpty(message="Password is mandatory")
    @NotBlank(message="Password is mandatory")
    @Size(min = 8,message="password should be at least 8 characters")
    private String password;
}